requirements.txt
python-telegram-bot[job-queue]==20.7
aiohttp>=3.8.0
pytz>=2023.3
geopy==2.4.1
